Allow changing dehydrated BASEDIR from env variable
This is to allow it to be placed in a volume that the user mounts into
the container, usually in /certs.
Introduce a new env variable DEHYDRATED_BASEDIR.
If this is unset, the behavior will stay the same as before:
`BASEDIR=/dehydrated`. This is for backwards compatibility.

`DEHYDRATED_BASEDIR` and `CERT_DIR` interact:
  * default (<=0.4.0) if `DEHYDRATED_BASEDIR` is not set:
    * `DEHYDRATED_BASEDIR="/dehydrated" CERT_DIR="/certs"`
    * `/dehydrated` is not stored in persistent storage and will be deleted with the container (including the letsencrypt account)
  * recommended (>0.4.0):
    * set `DEHYDRATED_BASEDIR="/certs/dehydrated"`. This will set `CERT_DIR="$DEHYDRATED_BASEDIR/certs"`. This allows all the certificates and dehydrated state (most importantly, accounts) to be stored in one directory, usually a volume mounted at `/certs`.

Introduce a new variable, DEHYDRATED_WELLKNOWN, to set the directory
where dehydrated will place .well-known. This allows the user to point
it to the webserver document root and not have to create any symlinks.